Allow disabling Terrain Ruler to ignore temporarily difficult terrain
Otherwise, Drag Ruler always respects difficult terrain, even when the normal ruler doesn't. This hooks into the window variable that Terrain Ruler sets to do so. A probably unnecessary check also doubles down on making sure that `window.terrainRuler` is real, just in case somehow the module is loaded but that isn't. This may matter if that variable name is changed down the line but I doubt it will be.